create Quiz component
