+++ pages/index.js
@@ -1,43 +1,8 @@
 import data from "../src/data"
-import { Box, Button, Container, Typography } from "@mui/material"
-import Image from "next/image"
-import { useState } from "react"
+import { Box, Container, Typography } from "@mui/material"
+import Quiz from '/src/components/quiz'
 
 const Home = () => {
-  const [answerIsShown, setAnswerIsShown] = useState(false)
-  const [isRightAnswer, setIsRightAnswer] = useState(false)
-
-  const defaultButtonStyle = [
-    { variant: 'outlined', color: 'primary' },
-    { variant: 'outlined', color: 'primary' },
-    { variant: 'outlined', color: 'primary' },
-  ]
-  const [buttonStyle, setButtonStyle] = useState(defaultButtonStyle)
-
-  const id = data[0].id
-  const choices = data[0].choices
-  const answer = data[0].answer
-
-  const judgeAnswer = (choiceIndex) => {
-    if (answerIsShown) return
-    setAnswerIsShown(true)
-    setButtonStyleFromAnswer(choiceIndex)
-    if (choiceIndex === answer) {
-      console.log('Right Answer!')
-      setIsRightAnswer(true)
-    }
-  }
-
-  const setButtonStyleFromAnswer = (choiceIndex) => {
-    const buttonStyle = defaultButtonStyle
-    const buttonStyleUpdated = buttonStyle.map((style, index) => {
-      if (index === answer) return { variant: 'contained', color: 'primary' }
-      if (index === answer || choiceIndex === index) return { variant: 'contained', color: 'secondary' }
-      return style
-    })
-    setButtonStyle(buttonStyleUpdated)
-  }
-
   return (
     <>
       <Container maxWidth="sm">
@@ -46,53 +11,9 @@ const Home = () => {
             ガチで東京の人しか解けない！ #東京の難読地名クイズ
           </Typography>
         </Box>
-        <Box sx={{mb: 5}}>
-          <Typography fontWeight="bold">
-            { id }. この地名はなんて読む？
-          </Typography>
-          <Box sx={{
-            mt: 3,
-            display: "flex",
-            justifyContent: "center"
-          }}>
-            <Image src="/img/kuizy01.png" width="620" height="372" />
-          </Box>
-          <Box sx={{width: '100%'}}>
-            {choices.map((choice, index) => {
-              return (
-                <Box sx={{width: '100%', mb: 2}}>
-                  <Button
-                    variant={buttonStyle[index].variant}
-                    fullWidth={true}
-                    color={buttonStyle[index].color}
-                    sx={{
-                      justifyContent: 'start',
-                      py: 1,
-                      fontWeight: 'bold',
-                    }}
-                    onClick={() => judgeAnswer(index)}
-                  >
-                    {choice}
-                  </Button>
-                </Box>
-              )
-            })}
-            <Box borderRadius={2} p={2} sx={{backgroundColor: '#f5f5f5'}}>
-              {Boolean(answerIsShown && isRightAnswer)
-                &&
-                <Box>
-                  <Typography sx={{ fontWeight: 'bold', borderBottom: '3px solid #1976d2', display: 'block', width: 'fit-content', mb: 2 }}>正解！</Typography>
-                  <Typography>正解は「{ choices[answer] }」です！</Typography>
-                </Box>}
-              {Boolean(answerIsShown && !isRightAnswer)
-                &&
-                <Box>
-                  <Typography sx={{ fontWeight: 'bold', borderBottom: '3px solid #9c27b0', display: 'block', width: 'fit-content', mb: 2 }}>不正解！</Typography>
-                  <Typography>正解は「{ choices[answer] }」です！</Typography>
-                </Box>}
-            </Box>
-          </Box>
-        </Box>
+        {data.map((datum) => {
+          return <Quiz datum={datum} />
+        })}
       </Container>
     </>
   )
